move tag functions to libnfc module
