+++ pkg/readers/libnfc/tags/ndef.go
@@ -19,7 +19,7 @@ You should have received a copy of the GNU General Public License
 along with TapTo.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package tokens
+package tags
 
 import (
 	"bytes"
